requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LLM and AI
openai==1.3.7
ollama==0.1.6
httpx==0.25.2

# Voice processing
pvporcupine==3.0.0
openai-whisper==20231117
pyttsx3==2.90
pyaudio==0.2.14

# Microsoft Graph API
msal==1.25.0
requests==2.31.0

# Email (IMAP)
email-validator==2.1.0

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.23

# Background tasks
apscheduler==3.10.4

# Utilities
pyyaml==6.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

